requirements.txt
langgraph
langchain-core
google-genai
beautifulsoup4
requests
tiktoken
faiss-cpu
sentence-transformers
pydantic
python-dotenv
jinja2
weasyprint
streamlit
langchain
langchain-google-genai
langgraph_supervisor
openai
fastapi
uvicorn
dotenv
python-docx